requirements.txt
numpy
pandas
scikit-learn
opencv-python
scikit-image
tensorflow
joblib
pillow
streamlit
xgboost
